+++ database/migrations/2022_02_05_021531_create_admins_table.php
@@ -16,7 +16,7 @@ class CreateAdminsTable extends Migration
         Schema::create('admins', function (Blueprint $table) {
             $table->increments('id');
             $table->string('name');
-            $table->integer('nip')->unique();
+            $table->bigInteger('nip')->unique();
             $table->integer('id_user')->unsigned();
             $table->integer('status')->default(0);
             $table->timestamps();
